[TAN-8601] Extract the legacy custom page header into its own component, avoiding ternary construction
